Refactor GoalId out of scheduler server
